ref: restructure and cleanup
