+++ packages/web-sdk/src/transports/otlp/otel-payload/attributes/attributeUtils.ts
