+++ packages/web-sdk/src/transports/otlp/otel-payload/attributes/index.ts
@@ -1,3 +1,2 @@
-export type { Attribute, AttributeValue } from './types';
-
+export { Attribute, AttributeValue } from './types';
 export { toAttribute, toAttributeValue, isAttribute } from './attributeUtils';
