+++ packages/web-sdk/src/transports/otlp/otel-payload/index.ts
@@ -1,0 +1,1 @@
+export { OtelPayload } from './OtelPayload';
