+++ packages/web-sdk/src/transports/otlp/otel-payload/transform/getEventLogRecord.test.ts
@@ -1,17 +1,20 @@
-import { LogEvent, LogLevel, TransportItem, TransportItemType } from '@grafana/faro-core';
-import { getScopeLog } from './transfoms';
+import { EventEvent, TransportItem, TransportItemType } from '@grafana/faro-core';
+import { getScopeLog } from './transforms';
 
-const item: TransportItem<LogEvent> = {
-  type: TransportItemType.LOG,
+const item: TransportItem<EventEvent> = {
+  type: TransportItemType.EVENT,
   payload: {
-    context: {},
-    level: LogLevel.INFO,
-    message: 'Faro was initialized',
+    name: 'event-name',
+    domain: 'event-domain',
     timestamp: '2023-01-27T09:53:01.035Z',
+    attributes: {
+      eventAttribute1: 'one',
+      eventAttribute2: 'two',
+    },
     trace: {
       trace_id: 'trace-id',
       span_id: 'span-id',
-    },
+    } as const,
   },
   meta: {
     view: {
@@ -40,17 +43,13 @@ const item: TransportItem<LogEvent> = {
         userAttribute1: 'one',
         userAttribute2: 'two',
       },
-    },
-  },
-} as const;
+    } as const,
+  } as const,
+};
 
-const logLogRecordPayload = {
+const errorLogRecordPayload = {
   timeUnixNano: 1674813181035000000,
-  severityNumber: 10, // static value
-  severityText: 'INFO2', // static value
-  body: {
-    stringValue: 'Faro was initialized',
-  },
+
   attributes: [
     {
       key: 'view.name',
@@ -149,12 +148,41 @@ const logLogRecordPayload = {
         },
       },
     },
+    {
+      key: 'event.name',
+      value: { stringValue: 'event-name' },
+    },
+    {
+      key: 'event.domain',
+      value: { stringValue: 'event-domain' },
+    },
+    {
+      key: 'event.attributes',
+      value: {
+        kvlistValue: {
+          values: [
+            {
+              key: 'eventAttribute1',
+              value: {
+                stringValue: 'one',
+              },
+            },
+            {
+              key: 'eventAttribute2',
+              value: {
+                stringValue: 'two',
+              },
+            },
+          ],
+        },
+      },
+    },
   ],
 } as const;
 
-describe('getLogLogRecord', () => {
+describe('getErrorLogRecord', () => {
   it('Builds resource payload object for given transport item.', () => {
     const logLogRecord = getScopeLog(item).logRecords[0];
-    expect(logLogRecord).toMatchObject(logLogRecordPayload);
+    expect(logLogRecord).toMatchObject(errorLogRecordPayload);
   });
 });
