+++ packages/web-sdk/src/transports/otlp/otel-payload/transform/index.ts
@@ -1,0 +1,11 @@
+export {
+  ErrorLogRecordPayload,
+  EventLogRecordPayload,
+  LogLogRecordPayload,
+  LogTransportItem,
+  ResourceLogPayload,
+  ResourcePayload,
+  ScopeLog,
+} from './types';
+
+export { getScopeLog, getResourceLogPayload } from './transforms';
