--- packages/web-sdk/src/transports/otlp/transform/index.ts
@@ -1,1 +1,0 @@
-
